Updated naming to SearchSolr and SearchElastic
