+++ src/main/java/io/anserini/search/SearchElastic.java
@@ -16,22 +16,32 @@
 
 package io.anserini.search;
 
-import com.google.common.base.Splitter;
 import io.anserini.rerank.ScoredDocuments;
 import io.anserini.rerank.lib.ScoreTiesAdjusterReranker;
 import io.anserini.search.topicreader.TopicReader;
 import io.anserini.index.generator.TweetGenerator;
 import org.apache.commons.lang3.time.DurationFormatUtils;
+import org.apache.http.HttpHost;
+import org.apache.http.auth.AuthScope;
+import org.apache.http.auth.UsernamePasswordCredentials;
+import org.apache.http.client.CredentialsProvider;
+import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.lucene.search.Query;
-import org.apache.lucene.document.LongPoint;
-import org.apache.solr.client.solrj.SolrClient;
-import org.apache.solr.client.solrj.impl.CloudSolrClient;
-import org.apache.solr.client.solrj.SolrQuery;
-import org.apache.solr.client.solrj.response.QueryResponse;
-import org.apache.solr.common.SolrDocumentList;
-import org.apache.solr.client.solrj.SolrQuery.SortClause;
+import org.elasticsearch.action.search.SearchRequest;
+import org.elasticsearch.action.search.SearchResponse;
+import org.elasticsearch.index.query.QueryBuilders;
+import org.elasticsearch.index.query.QueryStringQueryBuilder;
+import org.elasticsearch.index.query.BoolQueryBuilder;
+import org.elasticsearch.index.query.RangeQueryBuilder;
+import org.elasticsearch.search.builder.SearchSourceBuilder;
+import org.elasticsearch.search.SearchHits;
+import org.elasticsearch.client.RequestOptions;
+import org.elasticsearch.client.RestClient;
+import org.elasticsearch.client.RestHighLevelClient;
+import org.elasticsearch.search.sort.FieldSortBuilder;
+import org.elasticsearch.search.sort.ScoreSortBuilder;
+import org.elasticsearch.search.sort.SortOrder;
 import org.kohsuke.args4j.CmdLineException;
 import org.kohsuke.args4j.CmdLineParser;
 import org.kohsuke.args4j.OptionHandlerFilter;
@@ -48,7 +58,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Locale;
 import java.util.Map;
-import java.util.Optional;
 import java.util.SortedMap;
 import java.util.TreeMap;
 import java.util.concurrent.TimeUnit;
@@ -58,12 +67,12 @@ import static io.anserini.index.generator.LuceneDocumentGenerator.FIELD_ID;
 /*
 * Entry point of the Retrieval.
  */
-public final class SearchSolrCollection implements Closeable {
+public final class SearchElastic implements Closeable {
 
   private static final Logger LOG = LogManager.getLogger(SearchCollection.class);
   private static final int TIMEOUT = 600 * 1000;
   private final Args args;
-  private SolrClient client;
+  private RestHighLevelClient client;
 
   public static final class Args {
 
@@ -78,14 +87,23 @@ public final class SearchSolrCollection implements Closeable {
     @Option(name = "-topicreader", required = true, usage = "define how to read the topic(query) file: one of [Trec|Webxml]")
     public String topicReader;
 
-    @Option(name = "-solr.index", usage = "the name of the index in Solr")
-    public String solrIndex = null;
+    @Option(name = "-es.index", usage = "the name of the index in Elasticsearch")
+    public String esIndex = null;
 
-    @Option(name = "-solr.zkUrl", usage = "the URL of Solr's ZooKeeper (comma separated list of using ensemble)")
-    public String zkUrl = null;
+    @Option(name = "-es.hostname", usage = "the name of Elasticsearch HTTP host")
+    public String esHostname = "localhost";
 
-    @Option(name = "-solr.zkChroot", usage = "the ZooKeeper chroot")
-    public String zkChroot = "/";
+    @Option(name = "-es.port", usage = "the port for Elasticsearch HTTP host")
+    public int esPort = 9200;
+
+    /**
+     * The user and password are defaulted to those pre-configured for docker-elk
+     */
+    @Option(name = "-es.user", usage = "the user of the ELK stack")
+    public String esUser = "elastic";
+
+    @Option(name = "-es.password", usage = "the password for the ELK stack")
+    public String esPassword = "changeme";
 
     // optional arguments
     @Option(name = "-topicfield", usage = "Which field of the query should be used, default \"title\"." +
@@ -104,13 +122,13 @@ public final class SearchSolrCollection implements Closeable {
 
   }
 
-  private final class SolrSearcherThread<K> extends Thread {
+  private final class ESSearcherThread<K> extends Thread {
 
     final private SortedMap<K, Map<String, String>> topics;
     final private String outputPath;
     final private String runTag;
 
-    private SolrSearcherThread(SortedMap<K, Map<String, String>> topics, String outputPath, String runTag){
+    private ESSearcherThread(SortedMap<K, Map<String, String>> topics, String outputPath, String runTag){
 
       this.topics = topics;
       this.runTag = runTag;
@@ -121,7 +139,7 @@ public final class SearchSolrCollection implements Closeable {
     @Override
     public void run() {
       try {
-        LOG.info("[Start] Retrieval with Solr collection: " + args.solrIndex);
+        LOG.info("[Start] Retrieval with Elasticsearch collection: " + args.esIndex);
         final long start = System.nanoTime();
         PrintWriter out = new PrintWriter(Files.newBufferedWriter(Paths.get(outputPath), StandardCharsets.US_ASCII));
 
@@ -130,9 +148,9 @@ public final class SearchSolrCollection implements Closeable {
           String queryString = entry.getValue().get(args.topicfield);
           ScoredDocuments docs;
           if (args.searchtweets) {
-            docs = searchSolrTweets(queryString, Long.parseLong(entry.getValue().get("time")));
+            docs = searchTweets(queryString, Long.parseLong(entry.getValue().get("time")));
           } else {
-            docs = searchSolr(queryString);
+            docs = search(queryString);
           }
 
           /**
@@ -160,15 +178,19 @@ public final class SearchSolrCollection implements Closeable {
     }
   }
 
-  public SearchSolrCollection(Args args) throws IOException {
+  public SearchElastic(Args args) {
     this.args = args;
-    LOG.info("Solr index: " + args.solrIndex);
-    LOG.info("Solr ZooKeeper URL: " + args.zkUrl);
-    this.client = new CloudSolrClient.Builder(Splitter.on(',')
-            .splitToList(args.zkUrl), Optional.of(args.zkChroot))
-            .withConnectionTimeout(TIMEOUT)
-            .withSocketTimeout(TIMEOUT)
-            .build();
+    LOG.info("Elasticsearch index: " + args.esIndex);
+    LOG.info("Elasticsearch hostname: " + args.esHostname);
+    LOG.info("Elasticsearch host port: " + args.esPort);
+
+    final CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+    credentialsProvider.setCredentials(AuthScope.ANY, new UsernamePasswordCredentials(args.esUser, args.esPassword));
+
+    this.client = new RestHighLevelClient(
+            RestClient.builder(new HttpHost(args.esHostname, args.esPort, "http"))
+                    .setHttpClientConfigCallback(builder -> builder.setDefaultCredentialsProvider(credentialsProvider))
+                    .setRequestConfigCallback(builder -> builder.setConnectTimeout(TIMEOUT).setSocketTimeout(TIMEOUT)));
   }
 
   @SuppressWarnings("unchecked")
@@ -189,63 +211,91 @@ public final class SearchSolrCollection implements Closeable {
       }
     }
   
-    final String runTag = args.runtag == null ? "Solrini" : args.runtag;
-    SolrSearcherThread<K> solrThread = new SolrSearcherThread<K>(topics, args.output, runTag);
-    solrThread.run();
+    final String runTag = args.runtag == null ? "Elastirini" : args.runtag;
+    ESSearcherThread<K> esThread = new ESSearcherThread<K>(topics, args.output, runTag);
+    esThread.run();
   }
 
-  public<K> ScoredDocuments searchSolr(String queryString){
+  public<K> ScoredDocuments search(String queryString){
 
-    SolrDocumentList results = null;
+    SearchHits results = null;
 
-    SolrQuery solrq = new SolrQuery();
-    solrq.set("df", "contents");
-    solrq.set("fl", "* score");
-    // Remove double quotes in query since they are special syntax in Solr query parser
-    solrq.setQuery(queryString.replace("\"", ""));
-    solrq.setRows(args.hits);
-    solrq.setSort(SortClause.desc("score"));
-    solrq.addSort(SortClause.asc(FIELD_ID));
+    String specials = "+-=&|><!(){}[]^\"~*?:\\/";
+
+    for (int i = 0; i < specials.length(); i++){
+      char c = specials.charAt(i);
+      queryString = queryString.replace(String.valueOf(c), " ");
+    }
+
+    QueryStringQueryBuilder query = QueryBuilders
+            .queryStringQuery(queryString)
+            .defaultField("contents")
+            .analyzer("english");
+
+    SearchRequest searchRequest = new SearchRequest(args.esIndex);
+    SearchSourceBuilder sourceBuilder = new SearchSourceBuilder();
+    sourceBuilder.query(query);
+    sourceBuilder.size(args.hits);
+    sourceBuilder.sort(new ScoreSortBuilder().order(SortOrder.DESC));
+    sourceBuilder.sort(new FieldSortBuilder(FIELD_ID).order(SortOrder.ASC));
+    searchRequest.source(sourceBuilder);
 
     try {
-      QueryResponse response = client.query(args.solrIndex, solrq);
-      results = response.getResults();
+      SearchResponse searchResponse = client.search(searchRequest, RequestOptions.DEFAULT);
+      results = searchResponse.getHits();
     } catch (Exception e) {
-      LOG.error("Exception during Solr query: ", e);
+      LOG.error("Exception during ES query: ", e);
     }
 
     ScoreTiesAdjusterReranker reranker = new ScoreTiesAdjusterReranker();
-    return reranker.rerank(ScoredDocuments.fromSolrDocs(results), null);
+    return reranker.rerank(ScoredDocuments.fromESDocs(results), null);
   }
 
-  public<K> ScoredDocuments searchSolrTweets(String queryString, long t){
+  public<K> ScoredDocuments searchTweets(String queryString, long t){
 
-    SolrDocumentList results = null;
+    SearchHits results = null;
 
-    SolrQuery solrq = new SolrQuery();
-    solrq.set("df", "contents");
-    solrq.set("fl", "* score");
-    // Remove double quotes in query since they are special syntax in Solr query parser
-    solrq.setQuery(queryString.replace("\"", ""));
-    solrq.setRows(args.hits);
-    solrq.setSort(SortClause.desc("score"));
-    solrq.addSort(SortClause.desc(TweetGenerator.StatusField.ID_LONG.name));
+    String specials = "+-=&|><!(){}[]^\"~*?:\\/";
+
+    for (int i = 0; i < specials.length(); i++){
+      char c = specials.charAt(i);
+      queryString = queryString.replace(String.valueOf(c), " ");
+    }
 
     // Do not consider the tweets with tweet ids that are beyond the queryTweetTime
     // <querytweettime> tag contains the timestamp of the query in terms of the
     // chronologically nearest tweet id within the corpus
-    Query filter = LongPoint.newRangeQuery(TweetGenerator.StatusField.ID_LONG.name, 0L, t);
-    solrq.set("fq", filter.toString());
+    RangeQueryBuilder queryTweetTime = QueryBuilders
+            .rangeQuery(TweetGenerator.StatusField.ID_LONG.name)
+            .from(0L)
+            .to(t);
+
+    QueryStringQueryBuilder queryTerms = QueryBuilders
+            .queryStringQuery(queryString)
+            .defaultField("contents")
+            .analyzer("english");
+
+    BoolQueryBuilder query = QueryBuilders.boolQuery()
+            .filter(queryTweetTime)
+            .should(queryTerms);
+
+    SearchRequest searchRequest = new SearchRequest(args.esIndex);
+    SearchSourceBuilder sourceBuilder = new SearchSourceBuilder();
+    sourceBuilder.query(query);
+    sourceBuilder.size(args.hits);
+    sourceBuilder.sort(new ScoreSortBuilder().order(SortOrder.DESC));
+    sourceBuilder.sort(new FieldSortBuilder(TweetGenerator.StatusField.ID_LONG.name).order(SortOrder.DESC));
+    searchRequest.source(sourceBuilder);
 
     try {
-      QueryResponse response = client.query(args.solrIndex, solrq);
-      results = response.getResults();
+      SearchResponse searchResponse = client.search(searchRequest, RequestOptions.DEFAULT);
+      results = searchResponse.getHits();
     } catch (Exception e) {
-      LOG.error("Exception during Solr query: ", e);
+      LOG.error("Exception during ES query: ", e);
     }
 
     ScoreTiesAdjusterReranker reranker = new ScoreTiesAdjusterReranker();
-    return reranker.rerank(ScoredDocuments.fromSolrDocs(results), null);
+    return reranker.rerank(ScoredDocuments.fromESDocs(results), null);
   }
 
   @Override
@@ -254,20 +304,20 @@ public final class SearchSolrCollection implements Closeable {
   }
 
   public static void main(String[] args) throws Exception {
-    Args searchSolrArgs = new Args();
-    CmdLineParser parser = new CmdLineParser(searchSolrArgs, ParserProperties.defaults().withUsageWidth(90));
+    Args searchElasticArgs = new Args();
+    CmdLineParser parser = new CmdLineParser(searchElasticArgs, ParserProperties.defaults().withUsageWidth(90));
 
     try {
       parser.parseArgument(args);
     } catch (CmdLineException e) {
       System.err.println(e.getMessage());
       parser.printUsage(System.err);
-      System.err.println("Example: SearchSolrCollection" + parser.printExample(OptionHandlerFilter.REQUIRED));
+      System.err.println("Example: SearchElastic" + parser.printExample(OptionHandlerFilter.REQUIRED));
       return;
     }
 
     final long start = System.nanoTime();
-    SearchSolrCollection searcher = new SearchSolrCollection(searchSolrArgs);
+    SearchElastic searcher = new SearchElastic(searchElasticArgs);
     searcher.runTopics();
     searcher.close();
     final long durationMillis = TimeUnit.MILLISECONDS.convert(System.nanoTime() - start, TimeUnit.NANOSECONDS);
